utils: rename `hubstorage` module to `zyte` and update it
- use `ScrapinhubClient` instead of the deprecated `HubstorageClient`
- add spider() method
