+++ spidermon/utils/zyte.py
@@ -1,29 +1,59 @@
 """
-Module to hold a singleton Scrapinghub client.
+Module to hold a singleton to Zyte's Scrapy Cloud client.
+
+It expects that "SHUB_JOBKEY" be present as environment variable, as
+well as either "SH_APIKEY" or "SHUB_JOBAUTH" to access Scrapy Cloud API.
 """
 import os
 
 from scrapinghub import ScrapinghubClient
 
-from .hubstorage import _Hubstorage
 
-
-class Client(_Hubstorage):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._spider_name = os.environ.get("SHUB_SPIDER")
+class Client:
+    def __init__(self):
+        self.available = "SHUB_JOBKEY" in os.environ
+        self._client = None
+        self._project = None
+        self._spider = None
+        self._job = None
+        if self.available:
+            self.job_key = os.environ["SHUB_JOBKEY"]
+            self.project_id, self.spider_id, self.job_id = map(
+                int, self.job_key.split("/")
+            )
+        else:
+            self.project_id = None
+            self.spider_id = None
+            self.job_id = None
 
     @property
     def client(self):
-        if self._client is None:
+        if not self._client:
             self._client = ScrapinghubClient()
         return self._client
 
     @property
+    def project(self):
+        if not self._project:
+            self._project = self.client.get_project(str(self.project_id))
+        return self._project
+
+    @property
     def spider(self):
         if not self._spider:
-            self._spider = self.project.spiders.get(self._spider_name)
-        return self._project
+            spider_name = self.job.metadata.get("spider")
+            self._spider = self.project.spiders.get(spider_name)
+        return self._spider
+
+    @property
+    def job(self):
+        if not self._job:
+            self._job = self.client.get_job(self.job_key)
+        return self._job
+
+    def close(self):
+        if self._client:
+            self._client.close()
 
 
 client = Client()
